Pull out LIBDN unit testing code into a separate file
